requirements-app.txt
kivy>=2.2.1
kivymd>=1.1.1
opencv-python>=4.8.0
numpy>=1.21.0
pillow>=10.0.0
pytesseract>=0.3.10
android>=1.0.0
plyer>=2.1.0
jnius>=1.1.0
